+++ src/Service/Service.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Mediawiki\Api\Service;
+namespace Addwiki\MediaWikiApi\Service;
 
 use Mediawiki\Api\MediawikiApi;