+++ RELEASENOTES.md
@@ -3,10 +3,11 @@ Release Notes
 
 These are the release notes for [mediawiki-api](http://addwiki.readthedocs.io/projects/mediawiki-api/).
 
-# Version 0.7.4 (development)
+# Version 1.0.0 (development)
 
 * Fix PageGetter::getRevisionsFromResult undefined $revision['comment'] after an oversight/suppression of the edit summary.
 * Make `PagePurger::purge` purge `Page` objects that only have a title and no pageid
+* Move whole package to the `\Addwiki\WikiMediaApi` namespace, to conform to [PSR4](https://www.php-fig.org/psr/psr-4/)
 
 # Version 0.7.3 (14th January 2020)